+++ js/form/form.js
@@ -1,0 +1,7 @@
+import '../utils/utils.js';
+import '../server/server.js';
+import './validation/validation.js';
+import './tools/address-tool.js';
+import './tools/checkin-tool.js';
+import './tools/price-tool.js';
+import './tools/guest-tool.js';
