+++ js/map/map.js
@@ -1,0 +1,5 @@
+import '../utils/utils.js';
+import '../server/server.js';
+import './filter/filter.js';
+import './map-labels/map-labels.js';
+import './popup-baloon/popup-baloon.js';
