+++ js/server/server.js
@@ -1,0 +1,1 @@
+import '../utils/utils.js';
